Se añade Controller, Repositories, Service y editan models de comentario, descripcion, imagen , ubicacion y valoracion

+++ src/main/java/cl/gargolas/web/models/Imagen.java
@@ -36,7 +36,7 @@ import lombok.ToString;
 public class Imagen {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer idImagen;
+	private Long idImagen;
 	
 	@NotNull
 	private Blob imagen;
